% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{bear_df}
\alias{bear_df}
\title{bear_df}
\format{A data frame with 1000 rows and 3 variables}
\source{
\url{https://cran.r-project.org/web/packages/spatstat/vignettes/getstart.pdf) and [here](http://spatstat.org/resources/spatstatJSSpaper.pdf}
}
\description{
A dataset containing spatial information from the Scandinavian Bear Project
 The variables are as follows:
}
\keyword{datasets}
