% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{Data1}
\alias{Data1}
\title{Data1}
\format{A data frame with 65 rows and 6 variables}
\usage{
data(Data1)
}
\description{
A dataset contianing point data.
 The variables are as follows:
}
\details{
\itemize{
   \item x. x points based on a false origin and normalized to the unit-square  (0.87289164--0.88000000)
   \item y. y points based on a false origin and normalized to the unit-square  (0.13177294--0.85000000)
   \item VAR1. Variable assocaited with location used to adjust plot (388.55666--NA)
   \item VAR2. Variable associated with location used to adjust plot (571.07806--NA)
   \item VAR3. Variable assocaited with location used to adjust plot (634.7831--NA)
   \item observed. Variable states whether it was measured at that location (TRUE--FALSE)
}
}
\keyword{datasets}
