% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{df2_data6}
\alias{df2_data6}
\title{df2_data6}
\format{A data frame with 350 rows and 3 variables
source}
\description{
A dataset contianing simulated landscapes, one random, one with a strong systemic pattern
 The variables are as follows:
}
\keyword{datasets}
