% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{Fast_Food}
\alias{Fast_Food}
\title{Fast_Food
A dataset containing locations of fast food restaruants in Toronto (data is from 2008).
 The variables are as follows:}
\format{A data frame with 614 rows and 3 variables}
\description{
\itemize{
   \item x. list of x coordinates (\-79.31405--\-79.38232)
   \item y. list of y coordinates (43.82118--43.67041)
   \item class. Four sets of food offered at restaurants(Chicken, Hamburger, Pizza, Sub)
}
}
\keyword{datasets}
