% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{Gas_Stands}
\alias{Gas_Stands}
\title{Gas stands in the Toronto area}
\format{A data frame with 345 rows ad 2 variables}
\source{
Extracted from Toronto Business Points.RData
}
\usage{
data(Gas_Stands)
}
\description{
A simple feature (sf) dataframe containing the locations of gas stands in Toronto (data are from 2008). The projection is: EPSG: 26917 proj4string: "+proj=utm +zone=17 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs".
}
\keyword{datasets}
