% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{package}
\name{geog4ga3}
\alias{geog4ga3}
\alias{geog4ga3-package}
\title{geog4ga3}
\description{
A package with datasets and functions used in the course 4GA3 Applied Spatial Statistics offered by the School of Greography and Earth Sciences at McMaster University.
}
\author{
Antonio Paez
}
