% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{Hamilton_CT}
\alias{Hamilton_CT}
\title{Hamilton Census Tracts with select demographic information from the 2011 Census of Canada.}
\format{
A SpatialPolygonsDataDrame with 188 rows and 20 variables
}
\source{
\url{https://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/bound-limit-2011-eng.cfm}
}
\usage{
data(Hamilton_CT)
}
\description{
A simple feature (sf) dataframe of Hamilton Census Metropolitan Area (CMA) Census Tracts (CTs) with demographic information from the 2011 Census. The projection is: EPSG: 26917 proj4string: "+proj=utm +zone=17 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs"
}
\details{
\itemize{
   \item ID. Unique Identifier of Area
   \item AREA. Area of CT in sq.km.
   \item TRACT. Unique Identifier of Census Tract
   \item POPULATION. 2011 Population count in Census Tract
   \item POP_DENSITY. Population density in Census Tract
   \item AGE_LESS_20. Population aged less than 20 years old
   \item AGE_20_TO_24. Population aged 20 to 24 years old
   \item AGE_25_TO_29. Population aged 25 to 29 years old
   \item AGE_30_TO_34. Population aged 30 to 34 years old
   \item AGE_34_TO_39. Population aged 35 to 39 years old
   \item AGE_40_TO_44. Population aged 40 to 44 years old
   \item AGE_45_TO_49. Population aged 45 to 49 years old
   \item AGE_50_TO_54. Population aged 50 to 54 years old
   \item AGE_55_TO_59. Population aged 54 to 59 years old
   \item AGE_60_TO_64. Population aged 60 to 64 years old
   \item AGE_65_TO_69. Population aged 65 to 69 years old
   \item AGE_70_TO_74. Population aged 70 to 74 years old
   \item AGE_75_TO_79. Population aged 75 to 79 years old
   \item AGE_80_TO_84. Population aged 80 to 84 years old
   \item AGE_MORE_85. Population aged 85 years and older
}
}
\keyword{spatialpolygondataframe}
