% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpointmean.R
\name{kpointmean}
\alias{kpointmean}
\title{A funcion that implements k-point mean interpolation}
\usage{
kpointmean(source_xy, z, target_xy, k, longlat = FALSE)
}
\value{
A SpatialPolygonsDataFrame object.
}
\description{
This function calculates the spatial mean based on the k-nearest neighbors of a focal point or set of focal points.
}
\examples{
# Interpolate
target_xy = expand.grid(x = seq(0.5, 259.5, 2.2), y = seq(0.5, 299.5, 2.2))
source_xy = cbind(x = Walker_Lake$X, y = Walker_Lake$Y)
}
\keyword{spatial}
