% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localmoran.map.R
\name{localmoran.map}
\alias{localmoran.map}
\title{A funcion for creating an interactive map of local Moran's I statistics}
\usage{
localmoran.map(spat_pol = spat_pol, listw = listw, VAR = VAR,
  ID = ID)
}
\arguments{
\item{spat_pol, }{listw, VAR, ID}
}
\description{
This function obtains the m-surroundings by selecting the m-1 nearest neighbors of each observation, allowing for a degree of overlap of s.
}
\examples{
# Create a map of local Moran's I statistics for population density
localmoran.map(Hamilton_CT, Hamilton_CT.w, Hamilton_CT$POP_DENSIT, Hamilton_CT$TRACT)
}
\keyword{autocorrelation}
