% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{missing_df}
\alias{missing_df}
\title{missing_df}
\format{
A data frame with 65 rows and 6 variables
}
\source{
Data simulated by A. Paez
}
\usage{
data(missing_df)
}
\description{
A dataframe containing a set of observations, including some with missing variables.
 The variables are as follows:
}
\details{
\itemize{
   \item x. x points based on a false origin and normalized to the unit-square  (0.01698944--0.95718825)
   \item y. y points based on a false origin and normalized to the unit-square  (0.01003585--0.98715473)
   \item VAR1. Genric variable - think of it as housing prices or concentrations in ppb of some contaminant (50.00000--1050.00000)
   \item VAR2. Genric variable - think of it as housing prices or concentrations in ppb of some contaminant (50.00000--1050.00000)
   \item VAR3. Genric variable - think of it as housing prices or concentrations in ppb of some contaminant (50.00000--1050.00000)
   \item Observed. factor variable states whether the variables were measured for a location: if the status is "FALSE", the values of the variables are missing (TRUE--FALSE)
}
}
\keyword{datasets}
