% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{PointPattern2}
\alias{PointPattern2}
\title{PointPattern2}
\format{A data frame with 60 rows and 2 variables}
\description{
A dataset containing locations of some generic events.
 The variables are as follows:
}
\details{
\itemize{
   \item x. x points based on a false origin and normalized to the unit-square (0.21511760--0.13941605)
   \item y. y points based on a false origin and normalized to the unit-square (0.23998400--0.26875381)
}
}
\keyword{datasets}
