% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{PointPattern3}
\alias{PointPattern3}
\title{PointPattern3}
\format{A data frame with 60 rows and 2 variables}
\description{
A dataset containing locations of some generic events.
 The variables are as follows:
}
\details{
\itemize{
   \item x. x points based on a false origin and normalized to the unit-square (0.33453032--0.04994001)
   \item y. y points based on a false origin and normalized to the unit-square (0.31238907--0.32612196)
}
}
\keyword{datasets}
