% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{pp1_df}
\alias{pp1_df}
\title{pp1_df}
\format{A data frame with 72 rows and 3 variables}
\source{
Extracted from Data4.RData
}
\usage{
data(pp1_df)
}
\description{
A dataset that includes a `spatstat` `ppp`-class object with two sets of spatial events, labeled as "Pattern 1" and "Pattern 2".
 The varirables are as follows:
}
\details{
\itemize{
   \item x. list of x-coordinates used to plot two sets of spatial events (0.04559819--0.95643341)
   \item y. list of y-coordinates used to plot two sets of spatial events (0.03408578--0.94492099)
   \item marks. Two sets of spatial events (Pattern 1, Pattern 2)
}
}
\keyword{datasets}
