% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{pp2_df}
\alias{pp2_df}
\title{pp2_df}
\format{A data frame with 81 rows and 2 variables
source Dr. Paez}
\description{
A dataset containing saptial events
 The variables are as follows:
}
\keyword{datasets}
