% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{pp5_df}
\alias{pp5_df}
\title{pp5_df}
\format{A data frame with 81 rows and 2 variables}
\source{
Extracted from Data5.RData
}
\usage{
data(pp5_df)
}
\description{
A dataset containing `spatstat` `ppp`-class objects used to represent spatial events.
 The variables are as follows:
}
\details{
\itemize{
   \item x. list of x-coordinates (0.16--0.88)
   \item y. list of y-coordinates (0.16--0.88)
}
}
\keyword{datasets}
