% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{snow_deaths}
\alias{snow_deaths}
\title{John Snow's London 1854 cholera outbreak deaths}
\format{A data frame with 250 rows and 5 variables}
\source{
\url{http://blog.rtwilson.com/?s=snow}
}
\usage{
data(snow_deaths)
}
\description{
A dataset containing the location and number of cholera deaths in Soho, London.
 The variables are as follows:
}
\details{
\itemize{
  \item long. longitude of address (\-0.1401--\-0.1329)
  \item lat. latitude of address (51.51--51.52)
  \item Id. Unique identifier of address
  \item Count. Number of deaths recorded at address (1--15)
}
}
\keyword{datasets}
