% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{Toronto}
\alias{Toronto}
\title{Toronto}
\format{A data frame with 5250 rows and 7 variables}
\source{
\url{https://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/bound-limit-2011-eng.cfm}
}
\usage{
data(Toronto)
}
\description{
A dataset containing the city boundary of Toronto.
 The variables are as follows:
}
\details{
\itemize{
   \item lat. longitude of address (\-79.63930--\-79.37075)
   \item long. latitude of address (43.74985--43.62359)
   \item order. Unique idenifier of address
   \item hole. (FALSE)
   \item price. (1--10)
   \item group. (5299933.1-5299933.10)
   \item GTA06. (5299933)
}
}
\keyword{datasets}
