% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{trips_by_mode}
\alias{trips_by_mode}
\title{trips_by_mode}
\format{An excel file with 270 rows and 20 variables}
\source{
\url{http://www12.statcan.gc.ca/census-recensement/index-eng.cfm}
}
\usage{
data(trips_by_mode)
}
\description{
An excel file with the number of trips by mode of transportation by Traffic Analysis Zone (TAZ), and other useful information from the 2011 census for the Hamilton, CMA, Canada.
}
\details{
\itemize{
   \item GTA06. identifier used for spatial joins (4050--6020)
   \item Cycle. list of Hamiltonians that cycle to work (0--623)
   \item Auto_driver. list of Hamiltonians that drive to work (0--17743)
   \item Auto_passenger. list of Hamiltonians that get a ride to work (0--4321)
   \item Walk. list of Hamiltonians that walk to work (0--1599)
   \item Population. population based on a unique spatial polygon (38.88097--12770.552)
   \item Worked_in_2010_Full-time. number of Hamiltonians that worked full-time in 2010 (0--5925.9434)
   \item Worked_in_2010_Part-time. number of Hamiltonians that worked part-time in 2010 (0--1661.16313)
   \item Worked_at_home. number of Hamiltonians that worked from home (0--559.97542)
   \item Pop_Density. population denisty based on a unique spatial polygon (26.20745--14232.5677)
   \item Median_Age. median age of Hamiltonians based on a unique spatial polygon (3.845238--56.85006)
   \item Family_Size_2. size of family based on unique a spatial polygon (7.250167--1489.0255)
   \item Family_Size_3. size of family based on unique a spatial polygon (3.237384--859.09030)
   \item Family_Size_4. size of family based on unique a spatial polygon (1.619751--1281.18323)
   \item Family_Size_5_more. size of family based on a unique spatial polygon (1.617209--387.37487)
   \item Median_income. median income based on unique spatial polygon (9.496379--52496.09)
   \item Average_income. average income based on unique spatial polygon (11.44593--81235.73)
   \item Employment_rate. average employment rate based on a unique spatial polygon (32.74746--76.69758)
   \item Unemployment)rate. average unemployment rate based on a unique polygon (0.001258--23.200001)
   \item Median_commuting_duration. median commuting duration based on a unique polygon (15.41049--30.59950)
}
}
\keyword{datasets}
