% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{Walker_Lake}
\alias{Walker_Lake}
\title{Walker_Lake}
\format{A dataset with 470 rows and 6 variables
source \url{https://books.google.ca/books?id=vC2dcXFLI3YC&dq=introduction+to+applied+geostatistics+isaaks+and+srivastava&hl=en&sa=X&ved=0ahUKEwiKg6_iyrXZAhUjp1kKHd_jAVcQ6AEIKTAA}}
\description{
A dataset with geocoded observations of a series of variables originally used for teahcing geostatistics in Isaaks and Srivastava;s [An Introduction to Geostatistics]
 The variables are as follows:
}
\keyword{datasets}
