% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{Wolfcamp Aquifer}
\alias{Wolfcamp Aquifer}
\title{Wolfcamp Aquifer}
\format{A dataset with 85 rows and 3 variables
source \url{https://rubenfcasal.github.io/npsp/reference/aquifer.html}}
\description{
A dataset of piezometric head (watertable pressure) observations of the Wolfcamp Aquifer inn Texas
 The variables are as follows:
}
\keyword{datasets}
