% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{Paez_Mart}
\alias{Paez_Mart}
\title{A set of points with proposed Paez Mart locations}
\format{A data frame with 395 rows and 3 variables}
\source{
Locations proposed by A. Paez
}
\usage{
data(Paez_Mart)
}
\description{
A simple feature (sf) dataframe containing planned locations of convenience stores in Toronto. The projection is: EPSG: 26917 proj4string: "+proj=utm +zone=17 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs".
}
\keyword{datasets}
