% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog4ga3.R
\docType{data}
\name{Toronto}
\alias{Toronto}
\title{Toronto}
\format{
sf object with 1 observation of 1 variable (only geometry)
}
\source{
\url{https://www12.statcan.gc.ca/census-recensement/2011/geo/bound-limit/bound-limit-2011-eng.cfm}
}
\usage{
data(Toronto)
}
\description{
A simple feature (sf) dataframe of Toronto City boundaries. The projection is: EPSG: 26917 proj4string: "+proj=utm +zone=17 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs".
}
\keyword{datasets}
